Moved tiles to the purchase tab
