+++ src/stacks-hierarchy/navigation-types.ts
@@ -39,8 +39,8 @@ type PaymentParams = {
 export type RootStackParamList = {
   NotFound: undefined;
   Root_OnboardingStack: undefined;
-  Root_TipsAndInformation: NavigatorScreenParams<TabNavigatorStackParams>;
-  Root_TicketAssistantStack: NavigatorScreenParams<TabNavigatorStackParams>;
+  Root_TipsAndInformation: undefined;
+  Root_TicketAssistantStack: undefined;
   Root_TabNavigatorStack: NavigatorScreenParams<TabNavigatorStackParams>;
   Root_LocationSearchByTextScreen: Root_LocationSearchByTextScreenParams;
   Root_LocationSearchByMapScreen: Root_LocationSearchByMapScreenParams;
